Placed event parsing code in event.rs.
